Added GNU GPLv3 license
I'm going against my usual inclinations and going with gpl for this. For
most of my projects, I either want no one else to share the code
(battlemap), or the code to be shared as widely as possible (dice.js).
This means proprietary in the first case, mit or bsd in the latter.

In the first case, I don't want to encourage others to take and improve the
code at all, so I keep it closed. In the latter case, I don't mind if the
code is divergent. Getting pr's would be nice, but I'm happier to have it
be used than to get updates.

I feel differently about a programming language. Improvements _must_ be
shared, else we risk vendor lock and stagnation. Bugs not getting fixed, or
security patches not being upstreamed. Thus, the gpl.

My little bit of researc indicates programs created using milang do not
need to adhere to gpl, which I see as a general good. However, if you
modified milang to make your program, and then distribute the modified
milang, we (well, I) should at least be able to see what was changed.