Refactored the ast to be more 'regular'.
All this because...I don't really remember why now. I think I was trying
to do a more complicated compile, but I was running into issues with the
ast being too irregular? And that ended up bumping to needing to load
type information from other modules, including those that are system
modules (ie, not written in milang).

Anyway, the system headers are now written in milang, allowing them to
adapt and change as needed.