Converted to using a tokenizer then lexer.
This has made the errors during parse / compile much easier to decifer.
While there's still some value in the `parse` module, it's being relagated
to much simpler tasks.

The `milang_p_token` module generates a list of well defined token types.
It does not grammer checking beyond 'is this something that exists somewhere
in the milang grammer?'.

The `milang_lex` module turns that list of tokens and checks for actual
sensible grammer. As it finds it, it builds up an ast using `milang_ast`
to privide the types.

With that done, we can now have the compiler run the type validation or
generate a header from the created ast.