Massive revamp of ideoloty, syntax, and arch.
First a change in ideology. Initially milang was going to end up looking
very much like elm with some erlang mixed in and arbitrary infix
notations. However, that turned out to have a _ton_ of edge cases for
naming and how things looked. I hated having 2 ways that functions looked.
Different name binding rules. Further, it didn't really help reduce the
bike shedding that can happen.

I was playing around with scratch, and realized that the main issue with it
was a clunky interface. If an ide was able to input blocks of visual code
through text, then perhaps a visual programming language would work.

So, milang the language is a save file format(s) that is edited by
milang the ide. The tools to support compiling milang are still cli
forward, and the laguage itself is designed to be editable by humans.
However, it is not designed to be written primarily that way. It is
designed to be readable, however.

This lead to a massive change in syntax. It lead to a much more regular
system that did end up with more boilerplate and some keywords added.
However, code looks regular. The body of a function and the body
of a class implementation / class default look the same. Many features
have been (possibly on temporarily) tossed out. For example, there is no
module aliases for now, nor any auto matic aliasing of types / functions
from imported modules either. The ability is still there through the "let"
and "alias" declarations, it's just verbose. We're not worried about ease
of typing, so we're doing it. Other things have been massively moved.
Espose lists are gone. Instead a spec, type, or class declaration can have
the 'expose' or 'expose all' keyword in front of it. This allows one to
see at a glance what is exposed, and means a ui could toggle exposed or
not and have the data written more efficiently.

Finally, this meant massive changes for the parser. And that was ending up
painful. So I ended up re-writing a massive portion of that, adding a ton
of modules to represent the ast data parts. All this mainly to keep as much
comments w/ appropriate ast parts as I could.

Hopefully I've documented enough here, there, and elsewhere to allow all
this to make sense to someone other than myself, including my future self.